fix: address the circular dependencies among TranslationContext and Streami18n (#2483)
