+++ src/context/TranslationContext.tsx
@@ -4,58 +4,22 @@ import calendar from 'dayjs/plugin/calendar';
 import localizedFormat from 'dayjs/plugin/localizedFormat';
 
 import { getDisplayName } from './utils/getDisplayName';
+import { defaultDateTimeParser, defaultTranslatorFunction } from '../i18n/utils';
 
 import type { TFunction } from 'i18next';
-import type { Moment } from 'moment-timezone';
 import type { TranslationLanguages } from 'stream-chat';
 
 import type { UnknownType } from '../types/types';
-import { defaultTranslatorFunction } from '../i18n';
+import type { TDateTimeParser } from '../i18n/types';
 
 Dayjs.extend(calendar);
 Dayjs.extend(localizedFormat);
-
-export type SupportedTranslations =
-  | 'de'
-  | 'en'
-  | 'es'
-  | 'fr'
-  | 'hi'
-  | 'it'
-  | 'ja'
-  | 'ko'
-  | 'nl'
-  | 'pt'
-  | 'ru'
-  | 'tr';
-
-export const isLanguageSupported = (language: string): language is SupportedTranslations => {
-  const translations = ['de', 'en', 'es', 'fr', 'hi', 'it', 'ja', 'ko', 'nl', 'pt', 'ru', 'tr'];
-  return translations.some((translation) => language === translation);
-};
-
-export const isDayOrMoment = (output: TDateTimeParserOutput): output is Dayjs.Dayjs | Moment =>
-  !!(output as Dayjs.Dayjs | Moment)?.isSame;
-
-export const isDate = (output: TDateTimeParserOutput): output is Date =>
-  !!(output as Date)?.getMonth;
-
-export const isNumberOrString = (output: TDateTimeParserOutput): output is number | string =>
-  typeof output === 'string' || typeof output === 'number';
-
-export type TDateTimeParserInput = string | number | Date;
-
-export type TDateTimeParserOutput = string | number | Date | Dayjs.Dayjs | Moment;
-
-export type TDateTimeParser = (input?: TDateTimeParserInput) => TDateTimeParserOutput;
 
 export type TranslationContextValue = {
   t: TFunction;
   tDateTimeParser: TDateTimeParser;
   userLanguage: TranslationLanguages;
 };
-
-export const defaultDateTimeParser = (input?: TDateTimeParserInput) => Dayjs(input);
 
 export const TranslationContext = React.createContext<TranslationContextValue>({
   t: defaultTranslatorFunction,
