+++ src/i18n/Streami18n.ts
@@ -7,15 +7,13 @@ import localeData from 'dayjs/plugin/localeData';
 import relativeTime from 'dayjs/plugin/relativeTime';
 import utc from 'dayjs/plugin/utc';
 import timezone from 'dayjs/plugin/timezone';
-import { predefinedFormatters } from './utils';
+import { defaultTranslatorFunction, predefinedFormatters } from './utils';
 
 import type momentTimezone from 'moment-timezone';
 import type { TranslationLanguages } from 'stream-chat';
 
-import type { CustomFormatters, PredefinedFormatters } from './utils';
-import type { TDateTimeParser } from '../context/TranslationContext';
-
 import type { UnknownType } from '../types/types';
+import type { CustomFormatters, PredefinedFormatters, TDateTimeParser } from './types';
 
 import {
   deTranslations,
@@ -417,8 +415,6 @@ const defaultStreami18nOptions = {
   language: 'en' as TranslationLanguages,
   logger: (message?: string) => console.warn(message),
 };
-
-export const defaultTranslatorFunction: TFunction = <tResult = string>(key: tResult) => key;
 
 export class Streami18n {
   i18nInstance = i18n.createInstance();
