+++ src/i18n/index.ts
@@ -1,3 +1,11 @@
 export * from './translations';
 export * from './Streami18n';
-export type { FormatterFactory, TimestampFormatterOptions } from './utils';
+export {
+  defaultDateTimeParser,
+  defaultTranslatorFunction,
+  isDate,
+  isDayOrMoment,
+  isLanguageSupported,
+  isNumberOrString,
+} from './utils';
+export * from './types';
