+++ src/i18n/types.ts
@@ -1,0 +1,53 @@
+import type { Streami18n } from './Streami18n';
+import Dayjs from 'dayjs';
+import type { Moment } from 'moment-timezone';
+import { MessageContextValue } from '../context';
+import type { TFunction } from 'i18next';
+
+export type FormatterFactory<V> = (
+  streamI18n: Streami18n,
+) => (value: V, lng: string | undefined, options: Record<string, unknown>) => string;
+
+export type TimestampFormatterOptions = {
+  /* If true, call the `Day.js` calendar function to get the date string to display (e.g. "Yesterday at 3:58 PM"). */
+  calendar?: boolean;
+  /* Object specifying date display formats for dates formatted with calendar extension. Active only if calendar prop enabled. */
+  calendarFormats?: Record<string, string>;
+  /* Overrides the default timestamp format if calendar is disabled. */
+  format?: string;
+};
+
+export type TDateTimeParserInput = string | number | Date;
+export type TDateTimeParserOutput = string | number | Date | Dayjs.Dayjs | Moment;
+export type TDateTimeParser = (input?: TDateTimeParserInput) => TDateTimeParserOutput;
+
+export type SupportedTranslations =
+  | 'de'
+  | 'en'
+  | 'es'
+  | 'fr'
+  | 'hi'
+  | 'it'
+  | 'ja'
+  | 'ko'
+  | 'nl'
+  | 'pt'
+  | 'ru'
+  | 'tr';
+
+export type DateFormatterOptions = TimestampFormatterOptions & {
+  formatDate?: MessageContextValue['formatDate'];
+  messageCreatedAt?: string | Date;
+  t?: TFunction;
+  tDateTimeParser?: TDateTimeParser;
+  timestampTranslationKey?: string;
+};
+
+// Here is any used, because we do not want to enforce any specific rules and
+// want to leave the type declaration to the integrator
+/* eslint-disable-next-line  @typescript-eslint/no-explicit-any */
+export type CustomFormatters = Record<string, FormatterFactory<any>>;
+
+export type PredefinedFormatters = {
+  timestampFormatter: FormatterFactory<string | Date>;
+};
